ref(encryption): Refactor encrypted field
