+++ src/sentry/db/models/fields/encryption/__init__.py
@@ -1,0 +1,4 @@
+from .encrypted_char_field import EncryptedCharField
+from .encrypted_json_field import EncryptedJSONField
+
+__all__ = ["EncryptedCharField", "EncryptedJSONField"]
